+++ app/src/main/java/com/example/splitbill/User.kt
@@ -1,0 +1,8 @@
+package com.example.splitbill
+
+data class User(
+    val id: Int,
+    val name: String,
+    var amount: Long,
+    val isCurrentUser: Boolean
+)
